Consolidated education sql and contracts.
